Extract OpenPGP card operation logic from SecurityTokenConnection class
